vm code

+++ inc/core/bin_tree.h
@@ -10,15 +10,15 @@ struct btnode
 
 struct bintree
 {
-	int size;
-	int depth;
+	i32 size;
+	i32 depth;
 	struct btnode* root;
 };
 
 
 void bt_fillnew(struct btnode* node);
-long bt_insert(struct bintree* t, struct btnode* parent, struct btnode* node);
-long bt_remove(struct bintree* t, struct btnode* node);
+i32 bt_insert(struct bintree* t, struct btnode* parent, struct btnode* node);
+i32 bt_remove(struct bintree* t, struct btnode* node);
 struct btnode* bt_sibling(struct btnode* node);
 
 #endif// __bin_tree_h__
