+++ inc/core/dlist.h
@@ -9,24 +9,24 @@ struct dlnode
 
 struct dlist
 {
-//	long size;
+//	i32 size;
 	struct dlnode head;
 	struct dlnode tail;
 };
 
-long lst_new(struct dlist* lst);
-long lst_clr(struct dlnode* node);
-long lst_empty(struct dlist* lst);
+i32 lst_new(struct dlist* lst);
+i32 lst_clr(struct dlnode* node);
+i32 lst_empty(struct dlist* lst);
 
 
-long lst_insert_before(struct dlist* lst, struct dlnode* suc, struct dlnode* node);
-long lst_insert_after(struct dlist* lst, struct dlnode* prv, struct dlnode* node);
+i32 lst_insert_before(struct dlist* lst, struct dlnode* suc, struct dlnode* node);
+i32 lst_insert_after(struct dlist* lst, struct dlnode* prv, struct dlnode* node);
 
-long lst_remove(struct dlist* lst, struct dlnode* node);
-long lst_remove_node(struct dlnode* node);
+i32 lst_remove(struct dlist* lst, struct dlnode* node);
+i32 lst_remove_node(struct dlnode* node);
 
-long lst_push_back(struct dlist* lst, struct dlnode* node);
-long lst_push_front(struct dlist* lst, struct dlnode* node);
+i32 lst_push_back(struct dlist* lst, struct dlnode* node);
+i32 lst_push_front(struct dlist* lst, struct dlnode* node);
 
 struct dlnode* lst_pop_back(struct dlist* lst);
 struct dlnode* lst_pop_front(struct dlist* lst);
@@ -34,6 +34,6 @@ struct dlnode* lst_pop_front(struct dlist* lst);
 struct dlnode* lst_first(struct dlist* lst);
 struct dlnode* lst_last(struct dlist* lst);
 
-long lst_check(struct dlist* lst);
+i32 lst_check(struct dlist* lst);
 
 #endif
