+++ inc/core/pgpool.h
@@ -14,11 +14,11 @@ struct pgpool* pgp_create(void* addr, struct mm_config* cfg);
 struct pgpool* pgp_load(void* addr);
 void pgp_destroy(struct pgpool* up);
 
-void* pgp_alloc(struct pgpool* up, unsigned long size);
-long pgp_free(struct pgpool* up, void* p);
+void* pgp_alloc(struct pgpool* up, u64 size);
+i32 pgp_free(struct pgpool* up, void* p);
 
 
-long pgp_check(struct pgpool* up);
+i32 pgp_check(struct pgpool* up);
 
 #endif
