+++ inc/core/rbtree.h
@@ -21,7 +21,7 @@ typedef void (*order_function)(struct rbnode* n);
 // if key = n2->key, returns 0 
 // if key > n2->key, returns 1
 //
-typedef long (*compare_function)(void* key, struct rbnode* n);
+typedef i32 (*compare_function)(void* key, struct rbnode* n);
 
 struct rbtree
 {
@@ -37,7 +37,7 @@ struct rbnode* rb_parent(struct rbnode* node);
 struct rbnode* rb_sibling(struct rbnode* node);
 struct rbnode* rb_succ(struct rbnode* node);
 
-long rb_insert(struct rbtree* t, struct rbnode* node);
+i32 rb_insert(struct rbtree* t, struct rbnode* node);
 struct rbnode* rb_search(struct rbtree* t, void* key, struct rbnode** hot);
 struct rbnode* rb_remove(struct rbtree* t, void* key);
 void rb_remove_node(struct rbtree* t, struct rbnode* x);
