+++ inc/core/ringbuf.h
@@ -4,15 +4,15 @@
 struct ring_buf
 {
 	void* addr_begin;
-	unsigned long size;
+	u64 size;
 };
 
-struct ring_buf* rbuf_create(void* addr, long size);
-long rbuf_reset(struct ring_buf* rbuf);
-long rbuf_destroy(struct ring_buf* rbuf);
+struct ring_buf* rbuf_create(void* addr, i32 size);
+i32 rbuf_reset(struct ring_buf* rbuf);
+i32 rbuf_destroy(struct ring_buf* rbuf);
 
-long rbuf_write_block(struct ring_buf* rb, const void* data, long datalen);
-long rbuf_read_block(struct ring_buf* rb, void* buf, long buflen);
+i32 rbuf_write_block(struct ring_buf* rb, const void* data, i32 datalen);
+i32 rbuf_read_block(struct ring_buf* rb, void* buf, i32 buflen);
 
 
 #endif //__ring_buf_h__
